:fire: remove root container conflicting with mui's container + remove package react-toastify'
