+++ orgarif-client/package.json
@@ -19,7 +19,6 @@
     "@types/react-dom": "^17.0.5",
     "@types/react-input-mask": "^3.0.0",
     "@types/react-router-dom": "^5.1.7",
-    "@types/react-toastify": "^4.1.0",
     "es6-promise": "^4.2.8",
     "es6-shim": "^0.35.6",
     "es7-shim": "^6.0.0",
@@ -31,7 +30,6 @@
     "react-input-mask": "^2.0.4",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.3",
-    "react-toastify": "^8.0.3",
     "recoil": "^0.4.1",
     "typescript": "^4.3.2",
     "whatwg-fetch": "^3.6.2"
