--- orgarif-client/src/services/NotificationService.ts
@@ -1,18 +1,0 @@
-import { toast } from 'react-toastify';
-
-export type NotificationLevel = 'warning' | 'info';
-
-export class NotificationService {
-  public notifications = [] as [
-    string,
-    { message: string; level: NotificationLevel }
-  ][];
-
-  public displayNotification = (
-    message: string,
-    level: NotificationLevel = 'info'
-  ) => {
-    // TODO[error]
-    toast(message);
-  };
-}
